+++ baleen/baleen-core/src/test/java/uk/gov/dstl/baleen/testing/TestingUtils.java
@@ -14,10 +14,14 @@ public class TestingUtils {
 	}
 
 	public static boolean deleteDirectory(File dir) {
-		for (File f : dir.listFiles()) {
-			f.delete();
+		File[] files = dir.listFiles();
+		
+		if(files != null){
+			for (File f : files) {
+				f.delete();
+			}
+			dir.delete();
 		}
-		dir.delete();
 		return !dir.exists();
 	}
 }
